git-svn-id: file:///usr/gapps/silo/svn/branches/json-prototype@507 92378438-2262-0410-827d-edb41c1c42c8

+++ tests/ioperf_pdb.c
@@ -74,7 +74,7 @@ static int Write_pdb(void *buf, size_t nbytes)
     int status;
     char dsname[256];
     static int n = 0;
-    sprintf(dsname, "data_%07d(%d)", n++, nbytes/sizeof(double));
+    sprintf(dsname, "data_%07d(%d)", n++, (int) (nbytes/sizeof(double)));
     status = lite_PD_write(pdbfile, dsname, "double", buf);
     if (status) return nbytes;
     return 0;
